[DEV] - Adding tests for Event, Serie and Tag services

+++ src/main/java/com/example/web_project/api/service/EventService.java
@@ -42,6 +42,10 @@ public class EventService {
 
     public void deleteEvent(final long id_event)
     {
+        if(!eventRepository.existsById(id_event))
+        {
+            throw new EntityNotFoundException("Event not found with id: " + id_event);
+        }
         eventRepository.deleteById(id_event);
     }
